# Conflicts:
#	src/main/java/kg/edu/mathbilim/service/impl/UserServiceImpl.java
#	src/main/java/kg/edu/mathbilim/service/interfaces/UserService.java
#	src/main/resources/templates/profile/profile-page.ftlh

+++ src/main/java/kg/edu/mathbilim/service/interfaces/reference/UserTypeService.java
@@ -9,5 +9,5 @@ public interface UserTypeService {
 
     List<UserType> getAll();
 
-    UserType findById(Long id);
+    UserType findById(Integer id);
 }
